requirements.txt
# https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-win_amd64.whl
# https://download.pytorch.org/whl/cpu/torchvision-0.9.1%2Bcpu-cp39-cp39-win_amd64.whl
# albumentations
# iglovikov_helper_functions
# numpy
# pillow
streamlit
retinaface_pytorch  
stqdm